NLP - Sentiment Classification
The objective of this project is to build a text classification model that analyses the customer's sentiments 
based on their reviews in the IMDB database. The model uses a complex deep learning model to build an 
embedding layer followed by a classification algorithm to analyze the sentiment of the customers.